backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
openpyxl==3.1.5
SQLAlchemy==2.0.30
PyMySQL==1.1.1
python-multipart==0.0.9
openai==1.35.14
python-dotenv==1.0.1
